pyproject.toml
[tool.poetry]
name = "tennis-scoreboard"
version = "0.1.0"
description = ""
authors = ["ratmeow <rtm_bis@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.0"
alembic = "^1.14.0"
pytest = "^8.3.4"
waitress = "^3.0.2"
jinja2 = "^3.1.5"


[tool.poetry.group.dev.dependencies]
setuptools = "^75.6.0"
ruff = "^0.9.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100